Move cd() contextmanager out to utils file.
